feat(#690): move response.py from core/ to lib/ (tier-neutral utility)
HandlerResponse is a generic response wrapper used by backends/services,
not a kernel primitive. Fix 2x type: ignore[return-value] with cast().
