+++ src/nexus/backends/base_blob_connector.py
@@ -27,7 +27,7 @@ from typing import TYPE_CHECKING
 from nexus.backends.backend import Backend
 from nexus.contracts.exceptions import BackendError, NexusFileNotFoundError
 from nexus.core.hash_fast import hash_content
-from nexus.core.response import HandlerResponse
+from nexus.lib.response import HandlerResponse
 
 if TYPE_CHECKING:
     from nexus.contracts.types import OperationContext
