+++ tests/e2e/self_contained/isolation_helpers.py
@@ -65,40 +65,40 @@ class MockBackend:
     # ── content ops ──
 
     def write_content(self, content, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         h = hashlib.sha256(content).hexdigest()
         (self._content_dir / h).write_bytes(content)
         return HandlerResponse.ok(data=h, backend_name=self.name)
 
     def read_content(self, h, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         p = self._content_dir / h
         data = p.read_bytes() if p.exists() else b""
         return HandlerResponse.ok(data=data, backend_name=self.name)
 
     def delete_content(self, h, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         p = self._content_dir / h
         p.unlink(missing_ok=True)
         return HandlerResponse.ok(data=None, backend_name=self.name)
 
     def content_exists(self, h, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         return HandlerResponse.ok(data=(self._content_dir / h).exists(), backend_name=self.name)
 
     def get_content_size(self, h, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         p = self._content_dir / h
         size = p.stat().st_size if p.exists() else 0
         return HandlerResponse.ok(data=size, backend_name=self.name)
 
     def get_ref_count(self, h, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         return HandlerResponse.ok(
             data=1 if (self._content_dir / h).exists() else 0, backend_name=self.name
@@ -107,7 +107,7 @@ class MockBackend:
     # ── directory ops ──
 
     def mkdir(self, path, parents=False, exist_ok=False, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         dirs = self._load_dirs()
         dirs.add(path)
@@ -115,7 +115,7 @@ class MockBackend:
         return HandlerResponse.ok(data=None, backend_name=self.name)
 
     def rmdir(self, path, recursive=False, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         dirs = self._load_dirs()
         dirs.discard(path)
@@ -123,11 +123,11 @@ class MockBackend:
         return HandlerResponse.ok(data=None, backend_name=self.name)
 
     def is_directory(self, path, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         return HandlerResponse.ok(data=(path in self._load_dirs()), backend_name=self.name)
 
     def list_dir(self, path, context=None):
-        from nexus.core.response import HandlerResponse
+        from nexus.lib.response import HandlerResponse
 
         return HandlerResponse.ok(data=[], backend_name=self.name)
