+++ tests/unit/backends/test_batch_operations.py
@@ -15,7 +15,7 @@ from nexus.backends.backend import Backend
 from nexus.backends.base_blob_connector import BaseBlobStorageConnector
 from nexus.backends.cache_mixin import CacheConnectorMixin
 from nexus.contracts.types import OperationContext
-from nexus.core.response import HandlerResponse
+from nexus.lib.response import HandlerResponse
 
 
 class MockBlobConnector(BaseBlobStorageConnector, CacheConnectorMixin):
