+++ tests/unit/backends/test_encrypted_wrapper.py
@@ -24,7 +24,7 @@ from __future__ import annotations
 import pytest
 
 from nexus.core.protocols.describable import Describable
-from nexus.core.response import HandlerResponse
+from nexus.lib.response import HandlerResponse
 from tests.unit.backends.wrapper_test_helpers import make_leaf, make_storage_mock
 
 # ---------------------------------------------------------------------------
