+++ tests/unit/core/test_response.py
@@ -6,7 +6,7 @@ Tests the standardized response wrapper for backend operations.
 import pytest
 
 from nexus.contracts.exceptions import BackendError, ConflictError, NexusFileNotFoundError
-from nexus.core.response import HandlerResponse, ResponseType, timed_response
+from nexus.lib.response import HandlerResponse, ResponseType, timed_response
 
 
 class TestResponseType:
