+++ tests/unit/core/test_timed_response.py
@@ -15,7 +15,7 @@ import time
 import pytest
 
 from nexus.contracts.exceptions import BackendError, NexusFileNotFoundError
-from nexus.core.response import HandlerResponse, ResponseType, timed_response
+from nexus.lib.response import HandlerResponse, ResponseType, timed_response
 
 # === Fixtures ===
