+++ tests/unit/services/test_context_branch_merge.py
@@ -23,7 +23,7 @@ from sqlalchemy.orm import sessionmaker
 
 from nexus.contracts.exceptions import BranchConflictError, BranchStateError
 from nexus.contracts.workspace_manifest import ManifestEntry, WorkspaceManifest
-from nexus.core.response import HandlerResponse
+from nexus.lib.response import HandlerResponse
 from nexus.services.context_branch import ContextBranchService
 from nexus.storage.models._base import Base
 from nexus.storage.models.context_branch import ContextBranchModel
